*** empty log message ***
git-svn-id: http://code.jaspersoft.com/svn/repos/jasperreports/trunk/jasperreports@1347 76e25fc1-081a-0410-bb1b-c96bc70638b7

+++ src/net/sf/jasperreports/engine/fill/JRByteIncrementerFactory.java
@@ -255,7 +255,7 @@ class JRByteDistinctCountIncrementer extends JRAbstractExtendedIncrementer
 		DistinctCountHolder holder = 
 			(DistinctCountHolder)valueProvider.getValue(calculable.getHelperVariable(JRCalculable.HELPER_COUNT));
 		
-		return new Integer((int)holder.getCount());
+		return new Byte((byte)holder.getCount());
 	}
 	
 	public Object initialValue()
