+++ src/net/sf/jasperreports/engine/fill/JRLongIncrementerFactory.java
@@ -255,7 +255,7 @@ class JRLongDistinctCountIncrementer extends JRAbstractExtendedIncrementer
 		DistinctCountHolder holder = 
 			(DistinctCountHolder)valueProvider.getValue(calculable.getHelperVariable(JRCalculable.HELPER_COUNT));
 		
-		return new Integer((int)holder.getCount());
+		return new Long(holder.getCount());
 	}
 	
 	public Object initialValue()
