+++ src/net/sf/jasperreports/engine/fill/JRShortIncrementerFactory.java
@@ -255,7 +255,7 @@ class JRShortDistinctCountIncrementer extends JRAbstractExtendedIncrementer
 		DistinctCountHolder holder = 
 			(DistinctCountHolder)valueProvider.getValue(calculable.getHelperVariable(JRCalculable.HELPER_COUNT));
 		
-		return new Integer((int)holder.getCount());
+		return new Short((short)holder.getCount());
 	}
 	
 	public Object initialValue()
